+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added support to make JobService and ServiceInstance objects serializable when printed
 - Support Singularity compute in pipeline job
 - Added Feature Store, its dedicated classes and updated the docstrings, now available in public interface. The classes added are `FeatureStoreOperations, FeatureSetOperations, FeatureStoreEntityOperations` with properties classes specific to the new features.
+- Support additional_includes in command component
 
 ### Bugs Fixed